# graphdriver/copy: faster copy of hard links