@thaJeztah 2023-11-13
Some failure on Windows; looks like it already was restarted once (but haven't checked if it was the same failure); let me post it here in case it's related;

```
== Failed
=== FAIL: github.com/docker/docker/integration/container TestWindowsDevices/process/class/5B45201D-F2F2-4F3B-85BB-30FF1F953599_mounted (1.49s)
    devices_windows_test.go:107: assertion failed: error is not nil: Error response from daemon: failed to create task for container: failed to create shim task: hcs::CreateComputeSystem 979f82ae3a58a34a26c3c3116f0ccdb5558cb4ff8d7f2494def2165b81129da0: The parameter is incorrect.: unknown
```