@thaJeztah 2023-11-14
Out of curiousity; had a quick check why this was missing; The `ishardlink` was originally added in https://github.com/moby/moby/commit/bc5503f46bdc0816d90d99bd516afaa81717346a (https://github.com/moby/moby/pull/13067); but looks like the later optimization made in https://github.com/moby/moby/commit/b467f8b2ef21dc2239dcd136a29283ea6c3a0aee (https://github.com/moby/moby/pull/35537) did not take that into account.

Looking at this code, I'm now actually wondering if we should have that variable at all, or if we instead could just return early (which is what the `ishardlink` var does if it's set); perhaps that would make the code less brittle (keeping things together).