@robmry 2023-11-14
> Looking at this code, I'm now actually wondering if we should have that variable at all, or if we instead could just return early (which is what the `ishardlink` var does if it's set); perhaps that would make the code less brittle (keeping things together).

Minimal change seemed like the best plan for my first commit (!) ... but factoring the "switch" out into a function that returns something to say whether to copy the metadata would make the intention hard to miss.

(There might also be scope for speeding up regular-file copying by doing some of the i/o bound work in parallel on a pool of worker goroutines. But, speeding up the BusyBox-based integration tests seemed like a quick-win worth having, and I wasn't sure whether there was any driver to try to go further.)